Update QQ机器人服务定时重启脚本.bat
增加检查QQ进程，用于其他时间临时运行QQ时能定时关闭
重写部分代码
支持自定义QQ文件路径和QQ启动文件等配置项

+++ QQ机器人服务定时重启脚本.bat
@@ -1,37 +1,62 @@
 @echo off
 setlocal enabledelayedexpansion
 
-REM 执行脚本前关闭QQ进程
-:kill
-taskkill /IM QQ.exe /F
-goto loop
+REM 脚本配置
+set "QQ_Path=C:\Program Files\Tencent\QQNT"
+set "QQ_File=QQ.exe"
+set Time-based switch=0
+set start_time=1145
+set end_time=0130
 
-:loop
+:Process
 cls
+tasklist /fi "imagename eq QQ.exe" /nh | find /i "QQ.exe" >nul
+if %errorlevel% equ 0 (
+    if %Time-based switch% equ 1 (
+        call :time
+        call :run
+    ) else (
+        timeout /t 7200
+        taskkill /IM QQ.exe /F
+        goto :Process
+    )
+) else (
+    if %Time-based switch% equ 1 (
+        call :time
+        call :run
+    ) else (
+        echo 未检测到QQ运行
+        timeout /t 1 >nul
+        goto :Process
+    ) 
+)
+
+:time
 for /f "tokens=2 delims==" %%I in ('wmic os get localdatetime /value') do set datetime=%%I
 set current_hour=!datetime:~8,2!
 set current_minute=!datetime:~10,2!
-REM 设置QQ的启动时间与结束时间范围 (格式 HHMM)
-set start_time=1145
-set end_time=0130
 set /a current_time_minutes=(!current_hour!*60)+!current_minute!
 set /a start_time_minutes=(%start_time:~0,2%*60)+%start_time:~2,2%
 set /a end_time_minutes=(%end_time:~0,2%*60)+%end_time:~2,2%
-REM 支持跨天运行脚本
-if !end_time_minutes! lss !start_time_minutes! (
-    if !current_time_minutes! geq !start_time_minutes! ( goto run )
-    if !current_time_minutes! lss !end_time_minutes! ( goto run )
-) else (
-    if !current_time_minutes! geq !start_time_minutes! if !current_time_minutes! lss !end_time_minutes! ( goto run )
-)
-echo 检查时间是否符合条件
-timeout /t 60 >nul
-goto loop
+exit /b
 
 :run
-REM 再执行一次关闭QQ进程，防止在时间范围外已启动QQ再启动QQ冲突
+if !end_time_minutes! lss !start_time_minutes! (
+    if !current_time_minutes! geq !start_time_minutes! ( goto Loop )
+    if !current_time_minutes! lss !end_time_minutes! ( goto Loop )
+) else (
+    if !current_time_minutes! geq !start_time_minutes! if !current_time_minutes! lss !end_time_minutes! ( goto Loop )
+)
+echo 检查时间是否符合条件
+echo 当前时间%current_time_minutes%不符合条件
+timeout /t 60 >nul
+goto Process
+exit /b
+
+:Loop
 taskkill /IM QQ.exe /F
 timeout /t 2
-start "" /D "C:\Program Files\Tencent\QQNT" "QQ.exe"
+start "" /D "%QQ_Path%" "%QQ_File%"
 timeout /t 7200
-goto kill
+taskkill /IM QQ.exe /F
+goto Process
